Rewards
While logged in as a student, click the "Rewards" button.
Icons are placeholders (or are they???).
NOTE: This reads from the Coins element in XML, you'll have to manually
set the number in your XML to see how it properly works.
Milestones at 10, 25, 50, 100. After that, I have the math set up for
the "Up Next" to do every 100, but don't have any icons+labels to go
with.